Prevent a 0 UIScreen.main.scale value from causing problems. (#544)
Prevent a UIScreen.main.scale value of 0 from causing problems by returning a sane default (for most devices) of 3.0. This has been observed to occur in certain customer integrations.

We only use the value for specifying the density of the image we ask for Imgix to return, so as long as the value is not 0, then the correct layout for the image occurs (with the only potential issue of the resolution not perfectly matching the screen, which is almost certainly a non-issue).